+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ImGUIContext: Now debug UI is always drawn on top of all in-game UI elements.
 
+- Now current language setting is stored inside user settings .cfg file instead of project settings one.
+
 ### Fixed
 
 - (Issue #93) The screen is stretched when fullscreen mode is enabled via Alt+Enter hotkey issue was fixed.